// packages/components/src/monaco-editor/types/fileSystem.d.ts
export interface FileSystemAPI {
  readdir: ((path: string, options: "buffer" | {
    encoding: "buffer"
    withFileTypes?: false
  }) => Promise<Uint8Array[]>) & ((path: string, options?: {
    encoding?: BufferEncoding | null
    withFileTypes?: false
  } | BufferEncoding | null) => Promise<string[]>) & ((path: string, options: {
    encoding: "buffer"
    withFileTypes: true
  }) => Promise<DirEnt<Uint8Array>[]>) & ((path: string, options: {
    encoding?: BufferEncoding | null
    withFileTypes: true
  }) => Promise<DirEnt<string>[]>)
  readFile: ((path: string, encoding?: null) => Promise<Uint8Array>) & ((path: string, encoding: BufferEncoding) => Promise<string>)
  writeFile: (path: string, data: string | Uint8Array, options?: string | {
    encoding?: string | null
  } | null) => Promise<void>
  mkdir: ((path: string, options?: {
    recursive?: false
  }) => Promise<void>) & ((path: string, options: {
    recursive: true
  }) => Promise<string>)
  rm: (path: string, options?: {
    force?: boolean
    recursive?: boolean
  }) => Promise<void>
  rename: (oldPath: string, newPath: string) => Promise<void>
  watch?: ((filename: string, options?: FSWatchOptions, listener?: FSWatchCallback) => IFSWatcher) & ((filename: string, listener?: FSWatchCallback) => IFSWatcher)
}

export interface DirEnt<T> {
  name: T
  isFile: () => boolean
  isDirectory: () => boolean
}

export declare class WebContainer {
  /**
   * Gives access to the underlying file system.
   */
  fs: FileSystemAPI
  private static _instance
  private _tornDown
  private _unsubscribeFromTokenChangedListener
  /**
   * Spawns a process.
   *
   * @param command - The program to be executed.
   * @param args - The command-line arguments for the program.
   *
   * @example
   * ```
   * const install = await webcontainer.spawn('npm', ['i']);
   * ```
   */
  spawn(command: string, args: string[], options?: SpawnOptions): Promise<WebContainerProcess>
  /**
   * Spawns a process without command-line arguments.
   *
   * @param command - The program to be executed.
   *
   * @example
   * ```
   * const install = await webcontainer.spawn('yarn');
   * ```
   */
  spawn(command: string, options?: SpawnOptions): Promise<WebContainerProcess>
  /**
   * Exports the provided `path` in the format provided.
   *
   * @param path - The path to serialize.
   * @param options - Additional serialize options.
   *
   * @example
   * ```
   * const fileTree = await webcontainer.export('dist', { format: 'json' });
   * ```
   */
  export(path: string): Promise<FileSystemTree>
  export(path: string, options: ExportOptions & {
    format?: "json"
  }): Promise<FileSystemTree>
  export(path: string, options: ExportOptions): Promise<Uint8Array>
  /**
   * Listens for `port` events, which are emitted when a port is opened or closed by some process.
   */
  on(event: "port", listener: PortListener): Unsubscribe
  /**
   * Listens for `server-ready` events, emitted when a running server is listening for incoming
   * connections and ready to answer requests.
   */
  on(event: "server-ready", listener: ServerReadyListener): Unsubscribe
  /**
   * Listens for `preview-message` events, emitted when a preview sends a `PreviewMessage`.
   */
  on(event: "preview-message", listener: PreviewMessageListener): Unsubscribe
  /**
   * Listens for `error` events, emitted when an internal error is triggered.
   */
  on(event: "error", listener: ErrorListener): Unsubscribe
  /**
   * Listens for `xdg-open` events, emitted when the `xdg-open` command is invoked.
   */
  on(event: "xdg-open", listener: OpenListener): Unsubscribe
  /**
   * Listens for `code` events, emitted when the `code` command is invoked.
   */
  on(event: "code", listener: CodeListener): Unsubscribe
  /**
   * Mounts a tree of files into the filesystem. This can be specified as a tree object ({@link FileSystemTree})
   * or as a binary snapshot generated by [`@webcontainer/snapshot`](https://www.npmjs.com/package/@webcontainer/snapshot).
   *
   * @param snapshotOrTree - A tree of files, or a binary snapshot. Note that binary payloads will be transferred.
   * @param options.mountPoint - Specifies a nested path where the tree should be mounted.
   */
  mount(snapshotOrTree: FileSystemTree | Uint8Array | ArrayBuffer, options?: {
    mountPoint?: string
  }): Promise<void>
  /**
   * Set a custom script to be injected into all previews. When this function is called, every
   * future page reload will contain the provided script tag on all HTML responses.
   *
   * Note:
   *
   * When this function resolves, every preview reloaded _after_ will have the new script.
   * Existing preview have to be explicitely reloaded.
   *
   * To reload a preview you can use `reloadPreview`.
   *
   * @param scriptSrc Source for the script tag.
   * @param options Options to define which type of script this is.
   */
  setPreviewScript(scriptSrc: string, options?: PreviewScriptOptions): Promise<void>
  /**
   * The default value of the `PATH` environment variable for processes started through {@link spawn}.
   */
  get path(): string
  /**
   * The full path to the working directory (see {@link FileSystemAPI}).
   */
  get workdir(): string
  /**
   * Destroys the WebContainer instance, turning it unusable, and releases its resources. After this,
   * a new WebContainer instance can be obtained by calling {@link WebContainer.boot | `boot`}.
   *
   * All entities derived from this instance (e.g. processes, the file system, etc.) also become unusable
   * after calling this method.
   */
  teardown(): void
  /**
   * Boots a WebContainer. Only a single instance of WebContainer can be booted concurrently
   * (see {@link WebContainer.teardown | `teardown`}).
   *
   * Booting WebContainer is an expensive operation.
   */
  static boot(options?: BootOptions): Promise<WebContainer>
}

export interface FileSystemTree {
  [name: string]: DirectoryNode | FileNode | SymlinkNode
}
/**
 * Represents a directory, see {@link FileSystemTree}.
 */
export interface DirectoryNode {
  directory: FileSystemTree
}
/**
 * Represents a file, see {@link FileSystemTree}.
 */
export interface FileNode {
  file: {
    /**
     * The contents of the file, either as a UTF-8 string or as raw binary.
     */
    contents: string | Uint8Array
  }
}
/**
 * Represents a symlink, see {@link FileSystemTree}.
 */
export interface SymlinkNode {
  file: {
    /**
     * The target of the symlink.
     */
    symlink: string
  }
}